+++ edsl/jobs/interviews/InterviewTaskBuildingMixin.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 import asyncio
 import time
 import traceback
-from typing import Generator
+from typing import Generator, Union
 from edsl import CONFIG
 from edsl.exceptions import InterviewTimeoutError
 from edsl.data_transfer_models import AgentResponseDict
@@ -14,21 +14,31 @@ from edsl.jobs.buckets.ModelBuckets import ModelBuckets
 from edsl.jobs.interviews.interview_exception_tracking import InterviewExceptionEntry
 from edsl.jobs.interviews.retry_management import retry_strategy
 from edsl.jobs.tasks.task_status_enum import TaskStatus
-from edsl.jobs.tasks.TasksList import TasksList
 from edsl.jobs.tasks.QuestionTaskCreator import QuestionTaskCreator
-from edsl.data.Cache import Cache
+from edsl.agents.InvigilatorBase import InvigilatorBase
 
 TIMEOUT = float(CONFIG.get("EDSL_API_TIMEOUT"))
 
 
 class InterviewTaskBuildingMixin:
-    def _build_invigilators(self, debug: bool) -> Generator["Invigilator", None, None]:
-        """Create an invigilator for each question."""
-        for question in self.survey.questions:
-            yield self.get_invigilator(question=question, debug=debug)
+    def _build_invigilators(
+        self, debug: bool
+    ) -> Generator[InvigilatorBase, None, None]:
+        """Create an invigilator for each question.
 
-    def get_invigilator(self, question: QuestionBase, debug: bool) -> "Invigilator":
-        """Return an invigilator for the given question."""
+        :param debug: whether to use debug mode, in which case `InvigilatorDebug` is used.
+
+        An invigilator is responsible for answering a particular question in the survey.
+        """
+        for question in self.survey.questions:
+            yield self._get_invigilator(question=question, debug=debug)
+
+    def _get_invigilator(self, question: QuestionBase, debug: bool) -> "Invigilator":
+        """Return an invigilator for the given question.
+
+        :param question: the question to be answered
+        :param debug: whether to use debug mode, in which case `InvigilatorDebug` is used.
+        """
         invigilator = self.agent.create_invigilator(
             question=question,
             scenario=self.scenario,
@@ -43,23 +53,16 @@ class InterviewTaskBuildingMixin:
         """Return an invigilator for the given question."""
         return invigilator
 
-    @property
-    def dag(self) -> "DAG":
-        """Return the directed acyclic graph for the survey.
-
-        The DAG, or directed acyclic graph, is a dictionary that maps question names to their dependencies.
-        It is used to determine the order in which questions should be answered.
-        This reflects both agent 'memory' considerations and 'skip' logic.
-        The 'textify' parameter is set to True, so that the question names are returned as strings rather than integer indices.
-        """
-        return self.survey.dag(textify=True)
-
     def _build_question_tasks(
         self,
         debug: bool,
         model_buckets: ModelBuckets,
     ) -> list[asyncio.Task]:
-        """Create a task for each question, with dependencies on the questions that must be answered before this one can be answered."""
+        """Create a task for each question, with dependencies on the questions that must be answered before this one can be answered.
+
+        :param debug: whether to use debug mode, in which case `InvigilatorDebug` is used.
+        :param model_buckets: the model buckets used to track and control usage rates.
+        """
         tasks = []
         for question in self.survey.questions:
             tasks_that_must_be_completed_before = list(
@@ -82,13 +85,14 @@ class InterviewTaskBuildingMixin:
     ) -> Generator[asyncio.Task, None, None]:
         """Return the tasks that must be completed before the given question can be answered.
 
+        :param tasks: a list of tasks that have been created so far.
+        :param question: the question for which we are determining dependencies.
+
         If a question has no dependencies, this will be an empty list, [].
         """
         parents_of_focal_question = self.dag.get(question.question_name, [])
         for parent_question_name in parents_of_focal_question:
-            parent_index = self.to_index[parent_question_name]
-            parent_task = tasks[parent_index]
-            yield parent_task
+            yield tasks[self.to_index[parent_question_name]]
 
     def _create_question_task(
         self,
@@ -101,7 +105,13 @@ class InterviewTaskBuildingMixin:
     ) -> asyncio.Task:
         """Create a task that depends on the passed-in dependencies that are awaited before the task is run.
 
-        The task is created by a QuestionTaskCreator, which is responsible for creating the task and managing its dependencies.
+        :param question: the question to be answered. This is the question we are creating a task for.
+        :param tasks_that_must_be_completed_before: the tasks that must be completed before the focal task is run.
+        :param model_buckets: the model buckets used to track and control usage rates.
+        :param debug: whether to use debug mode, in which case `InvigilatorDebug` is used.
+        :param iteration: the iteration number for the interview.
+
+        The task is created by a `QuestionTaskCreator`, which is responsible for creating the task and managing its dependencies.
         It is passed a reference to the function that will be called to answer the question.
         It is passed a list "tasks_that_must_be_completed_before" that are awaited before the task is run.
         These are added as a dependency to the focal task.
@@ -123,7 +133,7 @@ class InterviewTaskBuildingMixin:
 
     def _get_estimated_request_tokens(self, question) -> float:
         """Estimate the number of tokens that will be required to run the focal task."""
-        invigilator = self.get_invigilator(question=question, debug=False)
+        invigilator = self._get_invigilator(question=question, debug=False)
         # TODO: There should be a way to get a more accurate estimate.
         combined_text = ""
         for prompt in invigilator.get_prompts().values():
@@ -148,68 +158,90 @@ class InterviewTaskBuildingMixin:
         This in turn calls the the passed-in agent's async_answer_question method, which returns a response dictionary.
         Note that is updates answers dictionary with the response.
         """
-        invigilator = self.get_invigilator(question, debug=debug)
+        invigilator = self._get_invigilator(question, debug=debug)
 
-        if skip := self._skip_this_question(question):
-            # print("Skipping question", question.question_name)
+        if self._skip_this_question(question):
             return invigilator.get_failed_task_result()
 
-        async def attempt_to_answer_question(invigilator):
-            try:
-                return await asyncio.wait_for(
-                    invigilator.async_answer_question(), timeout=TIMEOUT
-                )
-            except asyncio.TimeoutError as e:
-                exception_entry = InterviewExceptionEntry(
-                    exception=repr(e),
-                    time=time.time(),
-                    traceback=traceback.format_exc(),
-                )
-                if task:
-                    task.task_status = TaskStatus.FAILED
-                self.exceptions.add(question.question_name, exception_entry)
+        response: AgentResponseDict = await self._attempt_to_answer_question(
+            invigilator, task
+        )
 
-                raise InterviewTimeoutError(f"Task timed out after {TIMEOUT} seconds.")
-            except Exception as e:
-                exception_entry = InterviewExceptionEntry(
-                    exception=repr(e),
-                    time=time.time(),
-                    traceback=traceback.format_exc(),
-                )
-                if task:
-                    task.task_status = TaskStatus.FAILED
-                self.exceptions.add(question.question_name, exception_entry)
-                raise e
+        self._add_answer(response=response, question=question)
 
-        response: AgentResponseDict = await attempt_to_answer_question(invigilator)
-
-        self.answers.add_answer(response=response, question=question)
         # With the answer to the question, we can now cancel any skipped questions
-        # This is how the skip logic is implemented
         self._cancel_skipped_questions(question)
-
         return AgentResponseDict(**response)
 
+    def _add_answer(self, response: AgentResponseDict, question: QuestionBase) -> None:
+        """Add the answer to the answers dictionary.
+
+        :param response: the response to the question.
+        :param question: the question that was answered.
+        """
+        self.answers.add_answer(response=response, question=question)
+
     def _skip_this_question(self, current_question: QuestionBase) -> bool:
+        """Determine if the current question should be skipped.
+
+        :param current_question: the question to be answered.
+        """
         current_question_index = self.to_index[current_question.question_name]
 
         answers = self.answers | self.scenario | self.agent["traits"]
-
         skip = self.survey.rule_collection.skip_question_before_running(
             current_question_index, answers
         )
         return skip
 
+    async def _attempt_to_answer_question(
+        self, invigilator: InvigilatorBase, task: asyncio.Task
+    ) -> AgentResponseDict:
+        """Attempt to answer the question, and handle exceptions.
+
+        :param invigilator: the invigilator that will answer the question.
+        :param task: the task that is being run.
+        """
+        try:
+            return await asyncio.wait_for(
+                invigilator.async_answer_question(), timeout=TIMEOUT
+            )
+        except asyncio.TimeoutError as e:
+            exception_entry = InterviewExceptionEntry(
+                exception=repr(e),
+                time=time.time(),
+                traceback=traceback.format_exc(),
+            )
+            if task:
+                task.task_status = TaskStatus.FAILED
+            self.exceptions.add(invigilator.question.question_name, exception_entry)
+
+            raise InterviewTimeoutError(f"Task timed out after {TIMEOUT} seconds.")
+        except Exception as e:
+            exception_entry = InterviewExceptionEntry(
+                exception=repr(e),
+                time=time.time(),
+                traceback=traceback.format_exc(),
+            )
+            if task:
+                task.task_status = TaskStatus.FAILED
+            self.exceptions.add(invigilator.question.question_name, exception_entry)
+            raise e
+
     def _cancel_skipped_questions(self, current_question: QuestionBase) -> None:
         """Cancel the tasks for questions that are skipped.
+
+        :param current_question: the question that was just answered.
 
         It first determines the next question, given the current question and the current answers.
         If the next question is the end of the survey, it cancels all remaining tasks.
         If the next question is after the current question, it cancels all tasks between the current question and the next question.
         """
-        current_question_index = self.to_index[current_question.question_name]
+        current_question_index: int = self.to_index[current_question.question_name]
 
-        next_question = self.survey.rule_collection.next_question(
+        next_question: Union[
+            int, EndOfSurvey
+        ] = self.survey.rule_collection.next_question(
             q_now=current_question_index,
             answers=self.answers | self.scenario | self.agent["traits"],
         )
@@ -219,11 +251,7 @@ class InterviewTaskBuildingMixin:
         def cancel_between(start, end):
             """Cancel the tasks between the start and end indices."""
             for i in range(start, end):
-                task_to_cancel = self.tasks[i]
-                verbose = False
-                if verbose:
-                    print(f"Cancelling task {task_to_cancel.get_name()}")
-                task_to_cancel.cancel()
+                self.tasks[i].cancel()
 
         if next_question_index == EndOfSurvey:
             cancel_between(current_question_index + 1, len(self.survey.questions))