+++ CHANGELOG.md
@@ -309,7 +309,18 @@ We modified exception messages. If your survey run generates exceptions, run `re
 
 
 
-## [0.1.22] - 2024-05-TBD - These changes are not live
+## [0.1.22] - 2024-05-14
+
+### Added
+
+- New `Survey` method to export a survey to file. Usage: `generated_code = survey.code("example.py")`
+
+### Fixed
+
+- A bug in `Survey` method `add_skip_logic()`
+
+
+## [0.1.23] - 2024-05-18 - These changes are not live
 
 ### Added 
 
@@ -322,6 +333,18 @@ We modified exception messages. If your survey run generates exceptions, run `re
 ### Changed
 
 - [In progress] `QuestionMultipleChoice` is being modified allow non-responsive answers. Previously, an error was thrown if the agent did not select one of the given options. Details TBD.
+
+### Fixed
+
+### Deprecated
+
+### Removed
+
+## [0.1.24] - 2024-TBD
+
+### Added 
+
+### Changed
 
 ### Fixed